discord-bot/requirements.txt
discord.py==2.3.2
pinecone-client==3.0.0
openai==1.54.3
python-dotenv==1.0.0
aiohttp==3.9.1
httpx==0.27.0

